# perf: refactor AsyncTimer advice to Singleton