--- a/app/common/aop/AsyncTimer.ts
+++ b/app/common/aop/AsyncTimer.ts
@@ -4,30 +4,32 @@ import { Advice } from '@eggjs/tegg/aop';
 import { Inject } from '@eggjs/tegg';
 import type { EggLogger } from 'egg';
 
+const START = Symbol('AsyncTimer#start');
+const SUCCEED = Symbol('AsyncTimer#succeed');
+
 // auto print async function call performance timer log into logger
 @Advice()
 export class AsyncTimer implements IAdvice {
   @Inject()
   private readonly logger: EggLogger;
-  private start: number;
-  private succeed = true;
 
-  async beforeCall() {
-    this.start = performance.now();
+  async beforeCall(ctx: AdviceContext) {
+    ctx.set(START, performance.now());
+    ctx.set(SUCCEED, true);
   }
 
-  async afterThrow() {
-    this.succeed = false;
+  async afterThrow(ctx: AdviceContext) {
+    ctx.set(SUCCEED, false);
   }
 
   async afterFinally(ctx: AdviceContext) {
-    const ms = Math.floor((performance.now() - this.start) * 1000) / 1000;
+    const ms = Math.floor((performance.now() - ctx.get(START)!) * 1000) / 1000;
     this.logger.info(
       '[%s] [%s:%s|%s]',
       ms,
       ctx.that.constructor.name,
       ctx.method,
-      this.succeed ? 'T' : 'F'
+      ctx.get(SUCCEED)! ? 'T' : 'F'
     );
   }
 }