--- a/app/core/service/EventCorkerAdvice.ts
+++ b/app/core/service/EventCorkerAdvice.ts
@@ -1,9 +1,11 @@
 import type { ContextEventBus } from '@eggjs/tegg';
-import { Inject } from '@eggjs/tegg';
+import { Inject, ObjectInitType } from '@eggjs/tegg';
 import type { IAdvice } from '@eggjs/tegg/aop';
 import { Advice } from '@eggjs/tegg/aop';
 
-@Advice()
+@Advice({
+  initType: ObjectInitType.CONTEXT,
+})
 export class EventCorkAdvice implements IAdvice {
   @Inject()
   private eventBus: ContextEventBus;