<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Refactor**
  - Streamlined asynchronous operation tracking for improved performance monitoring.
  - Enhanced event handling configuration to provide more consistent and reliable system behavior.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->